Create PostMutationStore type
