+++ src/components/PostCard/Form.tsx
@@ -9,7 +9,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useUpdatePost } from '@/lib/hooks';
 import { postSchema } from '@/lib/schemas';
 import { usePostMutationStore } from '@/lib/stores';
-import { type PostSchema } from '@/lib/types';
+import { type PostMutationStore, type PostSchema } from '@/lib/types';
 
 import { Button } from '../Button';
 
@@ -37,7 +37,9 @@ const Form = (): ReactNode => {
   const { mutate: updatePost } = useUpdatePost();
 
   const setPostMutationData: (data: PostSchema) => void = usePostMutationStore(
-    (state) => state.setData
+    (state: PostMutationStore): ((data: PostSchema) => void) => {
+      return state.setData;
+    }
   );
 
   const onSubmit = (data: PostSchema): void => {
