+++ src/components/PostForm/PostForm.tsx
@@ -9,7 +9,7 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useCreatePost, useSignedInUser } from '@/lib/hooks';
 import { postSchema } from '@/lib/schemas';
 import { usePostMutationStore } from '@/lib/stores';
-import { type PostSchema } from '@/lib/types';
+import { type PostMutationStore, type PostSchema } from '@/lib/types';
 
 import { Button } from '../Button';
 
@@ -35,7 +35,12 @@ const PostForm = ({ className = '' }: Props): ReactNode => {
   });
 
   const { mutate: createPost } = useCreatePost();
-  const setPostMutationData = usePostMutationStore((state) => state.setData);
+
+  const setPostMutationData: (data: PostSchema) => void = usePostMutationStore(
+    (state: PostMutationStore): ((data: PostSchema) => void) => {
+      return state.setData;
+    }
+  );
 
   const onSubmit = (data: PostSchema): void => {
     createPost(data, {
