+++ src/lib/stores/usePostMutationStore.ts
@@ -1,13 +1,6 @@
 import { type StoreApi, type UseBoundStore, create } from 'zustand';
 
-import { type PostSchema } from '../types';
-
-type PostMutationStore = {
-  data: PostSchema | null;
-  id: number | undefined;
-  setData: (data: PostSchema) => void;
-  setId: (id: number | undefined) => void;
-};
+import { type PostMutationStore, type PostSchema } from '../types';
 
 const usePostMutationStore: UseBoundStore<StoreApi<PostMutationStore>> =
   create<PostMutationStore>((set) => {
