+++ src/lib/types/index.ts
@@ -6,6 +6,8 @@ export { type default as PostSchema } from './PostSchema';
 export { type default as ReactionSchema } from './ReactionSchema';
 export { type default as ViewSchema } from './ViewSchema';
 
+export { type default as PostMutationStore } from './PostMutationStore';
+
 export { type default as DummyJSONComment } from './DummyJSONComment';
 export { type default as DummyJSONPost } from './DummyJSONPost';
 export { type default as DummyJSONUser } from './DummyJSONUser';
